Run the entrypoint directly on exec
Relying on the Docker API to report the container status as running
does not guarantee that the entrypoint has finished running. The
entrypoint is a no-op if the user build already exists, so this will
only slow down the initial exec slightly. But hey, it's better than a
failure to launch!

Signed-off-by: Tyler Heucke <tyler.heucke@dell.com>